Fix bad merge of SearchActionComponentsTest
Tests on application has been removed when merging SearchActionComponentsMediumTest to SearchActionComponentsTest
