+++ server/sonar-server/src/test/java/org/sonar/server/issue/ws/SearchActionComponentsTest.java
@@ -20,6 +20,7 @@
 package org.sonar.server.issue.ws;
 
 import java.io.IOException;
+import java.util.Date;
 import java.util.List;
 import org.junit.After;
 import org.junit.Before;
@@ -59,56 +60,63 @@ import org.sonar.server.issue.workflow.FunctionExecutor;
 import org.sonar.server.issue.workflow.IssueWorkflow;
 import org.sonar.server.permission.index.AuthorizationTypeSupport;
 import org.sonar.server.permission.index.PermissionIndexer;
+import org.sonar.server.permission.index.PermissionIndexerTester;
 import org.sonar.server.tester.UserSessionRule;
 import org.sonar.server.view.index.ViewDoc;
 import org.sonar.server.view.index.ViewIndexDefinition;
 import org.sonar.server.view.index.ViewIndexer;
 import org.sonar.server.ws.WsActionTester;
 import org.sonar.server.ws.WsResponseCommonFormat;
-import org.sonarqube.ws.Issues;
+import org.sonarqube.ws.Issues.Issue;
 import org.sonarqube.ws.Issues.SearchWsResponse;
 import org.sonarqube.ws.client.issue.IssuesWsParameters;
 
 import static com.google.common.collect.Lists.newArrayList;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.sonar.api.utils.DateUtils.addDays;
 import static org.sonar.api.utils.DateUtils.parseDateTime;
 import static org.sonar.core.util.Uuids.UUID_EXAMPLE_01;
 import static org.sonar.core.util.Uuids.UUID_EXAMPLE_02;
 import static org.sonar.db.component.ComponentTesting.newFileDto;
 import static org.sonar.db.component.ComponentTesting.newModuleDto;
+import static org.sonar.db.component.ComponentTesting.newProjectCopy;
 import static org.sonar.db.component.SnapshotTesting.newAnalysis;
+import static org.sonarqube.ws.client.issue.IssuesWsParameters.PARAM_COMPONENT_KEYS;
+import static org.sonarqube.ws.client.issue.IssuesWsParameters.PARAM_PROJECT_KEYS;
+import static org.sonarqube.ws.client.issue.IssuesWsParameters.PARAM_SINCE_LEAK_PERIOD;
 
 public class SearchActionComponentsTest {
 
   @Rule
-  public UserSessionRule userSessionRule = UserSessionRule.standalone();
+  public UserSessionRule userSession = UserSessionRule.standalone();
   @Rule
-  public DbTester dbTester = DbTester.create();
+  public DbTester db = DbTester.create();
   @Rule
-  public EsTester esTester = new EsTester(new IssueIndexDefinition(new MapSettings().asConfig()), new ViewIndexDefinition(new MapSettings().asConfig()));
+  public EsTester es = new EsTester(new IssueIndexDefinition(new MapSettings().asConfig()), new ViewIndexDefinition(new MapSettings().asConfig()));
 
-  private DbClient db = dbTester.getDbClient();
-  private DbSession session = dbTester.getSession();
-  private IssueIndex issueIndex = new IssueIndex(esTester.client(), System2.INSTANCE, userSessionRule, new AuthorizationTypeSupport(userSessionRule));
-  private IssueIndexer issueIndexer = new IssueIndexer(esTester.client(), db, new IssueIteratorFactory(db));
-  private ViewIndexer viewIndexer = new ViewIndexer(db, esTester.client());
-  private IssueQueryFactory issueQueryFactory = new IssueQueryFactory(db, System2.INSTANCE, userSessionRule);
+  private DbClient dbClient = db.getDbClient();
+  private DbSession session = db.getSession();
+  private IssueIndex issueIndex = new IssueIndex(es.client(), System2.INSTANCE, userSession, new AuthorizationTypeSupport(userSession));
+  private IssueIndexer issueIndexer = new IssueIndexer(es.client(), dbClient, new IssueIteratorFactory(dbClient));
+  private ViewIndexer viewIndexer = new ViewIndexer(dbClient, es.client());
+  private IssueQueryFactory issueQueryFactory = new IssueQueryFactory(dbClient, System2.INSTANCE, userSession);
   private IssueFieldsSetter issueFieldsSetter = new IssueFieldsSetter();
   private IssueWorkflow issueWorkflow = new IssueWorkflow(new FunctionExecutor(issueFieldsSetter), issueFieldsSetter);
-  private SearchResponseLoader searchResponseLoader = new SearchResponseLoader(userSessionRule, db, new ActionFinder(userSessionRule), new TransitionService(userSessionRule, issueWorkflow));
+  private SearchResponseLoader searchResponseLoader = new SearchResponseLoader(userSession, dbClient, new ActionFinder(userSession), new TransitionService(userSession, issueWorkflow));
   private Languages languages = new Languages();
   private SearchResponseFormat searchResponseFormat = new SearchResponseFormat(new Durations(), new WsResponseCommonFormat(languages), languages, new AvatarResolverImpl());
-  private WsActionTester wsTester = new WsActionTester(new SearchAction(userSessionRule, issueIndex, issueQueryFactory, searchResponseLoader, searchResponseFormat));
+  private WsActionTester ws = new WsActionTester(new SearchAction(userSession, issueIndex, issueQueryFactory, searchResponseLoader, searchResponseFormat));
   private OrganizationDto defaultOrganization;
   private OrganizationDto otherOrganization1;
   private OrganizationDto otherOrganization2;
-  private StartupIndexer permissionIndexer = new PermissionIndexer(db, esTester.client(), issueIndexer);
+  private PermissionIndexerTester permissionIndexerTester = new PermissionIndexerTester(es, issueIndexer);
+  private StartupIndexer permissionIndexer = new PermissionIndexer(dbClient, es.client(), issueIndexer);
 
   @Before
   public void setUp() {
-    session = db.openSession(false);
-    OrganizationDao organizationDao = db.organizationDao();
-    this.defaultOrganization = dbTester.getDefaultOrganization();
+    session = dbClient.openSession(false);
+    OrganizationDao organizationDao = dbClient.organizationDao();
+    this.defaultOrganization = db.getDefaultOrganization();
     this.otherOrganization1 = OrganizationTesting.newOrganizationDto().setKey("my-org-1");
     this.otherOrganization2 = OrganizationTesting.newOrganizationDto().setKey("my-org-2");
     organizationDao.insert(session, this.otherOrganization1, false);
@@ -132,7 +140,7 @@ public class SearchActionComponentsTest {
       .setSeverity("MAJOR")
       .setIssueCreationDate(DateUtils.parseDateTime("2014-09-04T00:00:00+0100"))
       .setIssueUpdateDate(DateUtils.parseDateTime("2017-12-04T00:00:00+0100"));
-    db.issueDao().insert(session, issue);
+    dbClient.issueDao().insert(session, issue);
 
     ComponentDto project2 = insertComponent(ComponentTesting.newPublicProjectDto(otherOrganization2, "P2").setDbKey("PK2"));
     ComponentDto file2 = insertComponent(newFileDto(project2, null, "F2").setDbKey("FK2"));
@@ -142,12 +150,12 @@ public class SearchActionComponentsTest {
       .setSeverity("MAJOR")
       .setIssueCreationDate(DateUtils.parseDateTime("2014-09-04T00:00:00+0100"))
       .setIssueUpdateDate(DateUtils.parseDateTime("2017-12-04T00:00:00+0100"));
-    db.issueDao().insert(session, issue2);
+    dbClient.issueDao().insert(session, issue2);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .execute()
       .assertJson(this.getClass(), "issues_on_different_projects.json");
   }
@@ -160,15 +168,15 @@ public class SearchActionComponentsTest {
     RuleDto newRule = newRule();
     IssueDto issueInModule = IssueTesting.newDto(newRule, file, project).setKee("ISSUE_IN_MODULE");
     IssueDto issueInRootModule = IssueTesting.newDto(newRule, project, project).setKee("ISSUE_IN_ROOT_MODULE");
-    db.issueDao().insert(session, issueInModule, issueInRootModule);
+    dbClient.issueDao().insert(session, issueInModule, issueInRootModule);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    SearchWsResponse searchResponse = wsTester.newRequest().executeProtobuf(SearchWsResponse.class);
+    SearchWsResponse searchResponse = ws.newRequest().executeProtobuf(SearchWsResponse.class);
     assertThat(searchResponse.getIssuesCount()).isEqualTo(2);
 
-    for (Issues.Issue issue : searchResponse.getIssuesList()) {
+    for (Issue issue : searchResponse.getIssuesList()) {
       assertThat(issue.getProject()).isEqualTo("PK1");
       if (issue.getKey().equals("ISSUE_IN_MODULE")) {
         assertThat(issue.getSubProject()).isEqualTo("MK1");
@@ -183,27 +191,27 @@ public class SearchActionComponentsTest {
     ComponentDto project = insertComponent(ComponentTesting.newPublicProjectDto(otherOrganization1, "P1").setDbKey("PK1"));
     ComponentDto file = insertComponent(newFileDto(project, null, "F1").setDbKey("FK1"));
     IssueDto issue = IssueTesting.newDto(newRule(), file, project).setKee("82fd47d4-b650-4037-80bc-7b112bd4eac2");
-    db.issueDao().insert(session, issue);
+    dbClient.issueDao().insert(session, issue);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_PROJECT_UUIDS, project.uuid())
       .execute()
       .assertJson(this.getClass(), "search_by_project_uuid.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_PROJECT_UUIDS, "unknown")
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, project.uuid())
       .execute()
       .assertJson(this.getClass(), "search_by_project_uuid.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, "unknown")
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
@@ -213,7 +221,7 @@ public class SearchActionComponentsTest {
   public void search_since_leak_period_on_project() throws Exception {
     ComponentDto project = insertComponent(ComponentTesting.newPublicProjectDto(otherOrganization2, "P1").setDbKey("PK1"));
     ComponentDto file = insertComponent(newFileDto(project, null, "F1").setDbKey("FK1"));
-    db.snapshotDao().insert(session,
+    dbClient.snapshotDao().insert(session,
       newAnalysis(project)
         .setPeriodDate(parseDateTime("2015-09-03T00:00:00+0100").getTime()));
     RuleDto rule = newRule();
@@ -225,14 +233,14 @@ public class SearchActionComponentsTest {
       .setKee(UUID_EXAMPLE_02)
       .setIssueCreationDate(parseDateTime("2014-09-04T00:00:00+0100"))
       .setIssueUpdateDate(parseDateTime("2015-10-04T00:00:00+0100"));
-    db.issueDao().insert(session, issueAfterLeak, issueBeforeLeak);
+    dbClient.issueDao().insert(session, issueAfterLeak, issueBeforeLeak);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, project.uuid())
-      .setParam(IssuesWsParameters.PARAM_SINCE_LEAK_PERIOD, "true")
+      .setParam(PARAM_SINCE_LEAK_PERIOD, "true")
       .execute()
       .assertJson(this.getClass(), "search_since_leak_period.json");
   }
@@ -242,7 +250,7 @@ public class SearchActionComponentsTest {
     ComponentDto project = insertComponent(ComponentTesting.newPublicProjectDto(defaultOrganization, "P1").setDbKey("PK1"));
     ComponentDto module = insertComponent(newModuleDto(project));
     ComponentDto file = insertComponent(newFileDto(module, null, "F1").setDbKey("FK1"));
-    db.snapshotDao().insert(session,
+    dbClient.snapshotDao().insert(session,
       newAnalysis(project).setPeriodDate(parseDateTime("2015-09-03T00:00:00+0100").getTime()));
     RuleDto rule = newRule();
     IssueDto issueAfterLeak = IssueTesting.newDto(rule, file, project)
@@ -253,15 +261,15 @@ public class SearchActionComponentsTest {
       .setKee(UUID_EXAMPLE_02)
       .setIssueCreationDate(parseDateTime("2014-09-04T00:00:00+0100"))
       .setIssueUpdateDate(parseDateTime("2015-10-04T00:00:00+0100"));
-    db.issueDao().insert(session, issueAfterLeak, issueBeforeLeak);
+    dbClient.issueDao().insert(session, issueAfterLeak, issueBeforeLeak);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, project.uuid())
       .setParam(IssuesWsParameters.PARAM_FILE_UUIDS, file.uuid())
-      .setParam(IssuesWsParameters.PARAM_SINCE_LEAK_PERIOD, "true")
+      .setParam(PARAM_SINCE_LEAK_PERIOD, "true")
       .execute()
       .assertJson(this.getClass(), "search_since_leak_period.json");
   }
@@ -278,12 +286,12 @@ public class SearchActionComponentsTest {
     IssueDto issue1 = IssueTesting.newDto(rule, file1, project1).setKee("82fd47d4-b650-4037-80bc-7b112bd4eac2");
     IssueDto issue2 = IssueTesting.newDto(rule, file2, project2).setKee("2bd4eac2-b650-4037-80bc-7b1182fd47d4");
     IssueDto issue3 = IssueTesting.newDto(rule, file3, project3).setKee("7b1182fd-b650-4037-80bc-82fd47d4eac2");
-    db.issueDao().insert(session, issue1, issue2, issue3);
+    dbClient.issueDao().insert(session, issue1, issue2, issue3);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_PROJECT_UUIDS, project1.uuid())
       .setParam(WebService.Param.FACETS, "projectUuids")
       .execute()
@@ -295,27 +303,27 @@ public class SearchActionComponentsTest {
     ComponentDto project = insertComponent(ComponentTesting.newPublicProjectDto(defaultOrganization, "P1").setDbKey("PK1"));
     ComponentDto file = insertComponent(newFileDto(project, null, "F1").setDbKey("FK1"));
     IssueDto issue = IssueTesting.newDto(newRule(), file, project).setKee("82fd47d4-b650-4037-80bc-7b112bd4eac2");
-    db.issueDao().insert(session, issue);
+    dbClient.issueDao().insert(session, issue);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_FILE_UUIDS, file.uuid())
       .execute()
       .assertJson(this.getClass(), "search_by_file_uuid.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_FILE_UUIDS, "unknown")
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, file.uuid())
       .execute()
       .assertJson(this.getClass(), "search_by_file_uuid.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, "unknown")
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
@@ -329,17 +337,17 @@ public class SearchActionComponentsTest {
     RuleDto rule = newRule();
     IssueDto issueOnFile = IssueTesting.newDto(rule, file, project).setKee("82fd47d4-b650-4037-80bc-7b112bd4eac2");
     IssueDto issueOnTest = IssueTesting.newDto(rule, unitTest, project).setKee("2bd4eac2-b650-4037-80bc-7b1182fd47d4");
-    db.issueDao().insert(session, issueOnFile, issueOnTest);
+    dbClient.issueDao().insert(session, issueOnFile, issueOnTest);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENTS, file.getDbKey())
       .execute()
       .assertJson(this.getClass(), "search_by_file_key.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENTS, unitTest.getDbKey())
       .execute()
       .assertJson(this.getClass(), "search_by_test_key.json");
@@ -354,12 +362,12 @@ public class SearchActionComponentsTest {
     RuleDto newRule = newRule();
     IssueDto issue1 = IssueTesting.newDto(newRule, file1, project).setKee("82fd47d4-b650-4037-80bc-7b112bd4eac2");
     IssueDto issue2 = IssueTesting.newDto(newRule, file2, project).setKee("2bd4eac2-b650-4037-80bc-7b1182fd47d4");
-    db.issueDao().insert(session, issue1, issue2);
+    dbClient.issueDao().insert(session, issue1, issue2);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, project.uuid())
       .setParam(IssuesWsParameters.PARAM_FILE_UUIDS, file1.uuid() + "," + file3.uuid())
       .setParam(WebService.Param.FACETS, "fileUuids")
@@ -373,27 +381,27 @@ public class SearchActionComponentsTest {
     ComponentDto directory = insertComponent(ComponentTesting.newDirectory(project, "D1", "src/main/java/dir"));
     ComponentDto file = insertComponent(newFileDto(project, null, "F1").setDbKey("FK1").setPath(directory.path() + "/MyComponent.java"));
     IssueDto issue = IssueTesting.newDto(newRule(), file, project).setKee("82fd47d4-b650-4037-80bc-7b112bd4eac2");
-    db.issueDao().insert(session, issue);
+    dbClient.issueDao().insert(session, issue);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, directory.uuid())
       .execute()
       .assertJson(this.getClass(), "search_by_file_uuid.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, "unknown")
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_DIRECTORIES, "src/main/java/dir")
       .execute()
       .assertJson(this.getClass(), "search_by_file_uuid.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_DIRECTORIES, "src/main/java")
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
@@ -410,39 +418,39 @@ public class SearchActionComponentsTest {
     insertComponent(newFileDto(module2, directory2, "F2").setDbKey("FK2").setPath(directory2.path() + "/MyComponent.java"));
     RuleDto rule = newRule();
     IssueDto issue1 = IssueTesting.newDto(rule, file1, project).setKee("82fd47d4-b650-4037-80bc-7b112bd4eac2");
-    db.issueDao().insert(session, issue1);
+    dbClient.issueDao().insert(session, issue1);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, directory1.uuid())
       .execute()
       .assertJson(this.getClass(), "search_by_directory_uuid.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, directory2.uuid())
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_MODULE_UUIDS, module1.uuid())
       .setParam(IssuesWsParameters.PARAM_DIRECTORIES, "src/main/java/dir")
       .execute()
       .assertJson(this.getClass(), "search_by_directory_uuid.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_MODULE_UUIDS, module2.uuid())
       .setParam(IssuesWsParameters.PARAM_DIRECTORIES, "src/main/java/dir")
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_DIRECTORIES, "src/main/java/dir")
       .execute()
       .assertJson(this.getClass(), "search_by_directory_uuid.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_DIRECTORIES, "src/main/java")
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
@@ -460,12 +468,12 @@ public class SearchActionComponentsTest {
     RuleDto newRule = newRule();
     IssueDto issue1 = IssueTesting.newDto(newRule, file1, project).setKee("82fd47d4-b650-4037-80bc-7b112bd4eac2");
     IssueDto issue2 = IssueTesting.newDto(newRule, file2, project).setKee("2bd4eac2-b650-4037-80bc-7b1182fd47d4");
-    db.issueDao().insert(session, issue1, issue2);
+    dbClient.issueDao().insert(session, issue1, issue2);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, module.uuid())
       .setParam(IssuesWsParameters.PARAM_MODULE_UUIDS, subModule1.uuid() + "," + subModule3.uuid())
       .setParam(WebService.Param.FACETS, "moduleUuids")
@@ -479,13 +487,13 @@ public class SearchActionComponentsTest {
     ComponentDto directory = insertComponent(ComponentTesting.newDirectory(project, "D1", "src/main/java/dir"));
     ComponentDto file = insertComponent(newFileDto(project, directory, "F1").setDbKey("FK1").setPath(directory.path() + "/MyComponent.java"));
     IssueDto issue = IssueTesting.newDto(newRule(), file, project).setKee("82fd47d4-b650-4037-80bc-7b112bd4eac2");
-    db.issueDao().insert(session, issue);
+    dbClient.issueDao().insert(session, issue);
     session.commit();
     indexIssues();
     indexPermissions();
 
-    userSessionRule.logIn("john");
-    wsTester.newRequest()
+    userSession.logIn("john");
+    ws.newRequest()
       .setParam("resolved", "false")
       .setParam(WebService.Param.FACETS, "directories")
       .execute()
@@ -502,10 +510,10 @@ public class SearchActionComponentsTest {
 
     insertIssue(IssueTesting.newDto(newRule(), file, project).setKee("82fd47d4-b650-4037-80bc-7b112bd4eac2"));
 
-    userSessionRule.logIn("john")
+    userSession.logIn("john")
       .registerComponents(project, file, view);
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, view.uuid())
       .execute()
       .assertJson(this.getClass(), "search_by_view_uuid.json");
@@ -523,9 +531,9 @@ public class SearchActionComponentsTest {
     indexView(subView.uuid(), newArrayList(project.uuid()));
     indexPermissions();
 
-    userSessionRule.logIn("john")
+    userSession.logIn("john")
       .registerComponents(project, file, view, subView);
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, subView.uuid())
       .execute()
       .assertJson(this.getClass(), "search_by_view_uuid.json");
@@ -543,10 +551,10 @@ public class SearchActionComponentsTest {
     indexView(subView.uuid(), newArrayList(project.uuid()));
 
     // User has wrong permission on the view, no issue will be returned
-    userSessionRule.logIn("john")
+    userSession.logIn("john")
       .registerComponents(project, file, view, subView);
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_COMPONENT_UUIDS, subView.uuid())
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
@@ -561,21 +569,195 @@ public class SearchActionComponentsTest {
     IssueDto issue2 = IssueTesting.newDto(newRule, file, project).setAuthorLogin("luke@skywalker.name").setKee("82fd47d4-b650-4037-80bc-7b1182fd47d4");
     indexPermissions();
 
-    db.issueDao().insert(session, issue1, issue2);
+    dbClient.issueDao().insert(session, issue1, issue2);
     session.commit();
     indexIssues();
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_AUTHORS, "leia")
       .setParam(WebService.Param.FACETS, "authors")
       .execute()
       .assertJson(this.getClass(), "search_by_authors.json");
 
-    wsTester.newRequest()
+    ws.newRequest()
       .setParam(IssuesWsParameters.PARAM_AUTHORS, "unknown")
       .execute()
       .assertJson(this.getClass(), "no_issue.json");
+  }
+  @Test
+  public void search_by_application_key() throws Exception {
+    ComponentDto project1 = db.components().insertPublicProject();
+    ComponentDto project2 = db.components().insertPublicProject();
+    ComponentDto application = db.components().insertApplication(db.getDefaultOrganization());
+    db.components().insertComponents(newProjectCopy("PC1", project1, application));
+    db.components().insertComponents(newProjectCopy("PC2", project2, application));
+    IssueDto issue1 = db.issues().insertIssue(i -> i.setProject(project1));
+    IssueDto issue2 = db.issues().insertIssue(i -> i.setProject(project2));
+    userSession.registerComponents(application, project1, project2);
+    permissionIndexerTester.allowOnlyAnyone(project1);
+    permissionIndexerTester.allowOnlyAnyone(project2);
+    indexIssuesAndViews();
 
+    SearchWsResponse result = ws.newRequest()
+      .setParam(PARAM_COMPONENT_KEYS, application.getDbKey())
+      .executeProtobuf(SearchWsResponse.class);
+
+    assertThat(result.getIssuesList()).extracting(Issue::getKey)
+      .containsExactlyInAnyOrder(issue1.getKey(), issue2.getKey());
+  }
+
+  @Test
+  public void ignore_application_without_browse_permission() {
+    ComponentDto project = db.components().insertPublicProject();
+    ComponentDto application = db.components().insertApplication(db.getDefaultOrganization());
+    db.components().insertComponents(newProjectCopy("PC1", project, application));
+    db.issues().insertIssue(i -> i.setProject(project));
+    userSession.registerComponents(project);
+    permissionIndexerTester.allowOnlyAnyone(project);
+    indexIssuesAndViews();
+
+    SearchWsResponse result = ws.newRequest()
+      .setParam(PARAM_COMPONENT_KEYS, application.getDbKey())
+      .executeProtobuf(SearchWsResponse.class);
+
+    assertThat(result.getIssuesList()).isEmpty();
+  }
+
+  @Test
+  public void search_application_without_projects() {
+    ComponentDto project = db.components().insertPublicProject();
+    ComponentDto application = db.components().insertApplication(db.getDefaultOrganization());
+    db.issues().insertIssue(i -> i.setProject(project));
+    userSession.registerComponents(application, project);
+    permissionIndexerTester.allowOnlyAnyone(project);
+    indexIssuesAndViews();
+
+    SearchWsResponse result = ws.newRequest()
+      .setParam(PARAM_COMPONENT_KEYS, application.getDbKey())
+      .executeProtobuf(SearchWsResponse.class);
+
+    assertThat(result.getIssuesList()).isEmpty();
+  }
+
+  @Test
+  public void search_by_application_and_by_leak() throws Exception {
+    Date now = new Date();
+    ComponentDto application = db.components().insertApplication(db.getDefaultOrganization());
+    // Project 1
+    ComponentDto project1 = db.components().insertPublicProject();
+    db.components().insertSnapshot(project1, s -> s.setPeriodDate(addDays(now, -14).getTime()));
+    db.components().insertComponents(newProjectCopy("PC1", project1, application));
+    IssueDto project1Issue1 = db.issues().insertIssue(i -> i.setProject(project1).setIssueCreationDate(addDays(now, -10)));
+    IssueDto project1Issue2 = db.issues().insertIssue(i -> i.setProject(project1).setIssueCreationDate(addDays(now, -20)));
+    // Project 2
+    ComponentDto project2 = db.components().insertPublicProject();
+    db.components().insertSnapshot(project2, s -> s.setPeriodDate(addDays(now, -25).getTime()));
+    db.components().insertComponents(newProjectCopy("PC2", project2, application));
+    IssueDto project2Issue1 = db.issues().insertIssue(i -> i.setProject(project2).setIssueCreationDate(addDays(now, -15)));
+    IssueDto project2Issue2 = db.issues().insertIssue(i -> i.setProject(project2).setIssueCreationDate(addDays(now, -30)));
+    // Permissions and index
+    userSession.registerComponents(application, project1, project2);
+    permissionIndexerTester.allowOnlyAnyone(project1);
+    permissionIndexerTester.allowOnlyAnyone(project2);
+    indexIssuesAndViews();
+
+    SearchWsResponse result = ws.newRequest()
+      .setParam(PARAM_COMPONENT_KEYS, application.getDbKey())
+      .setParam(PARAM_SINCE_LEAK_PERIOD, "true")
+      .executeProtobuf(SearchWsResponse.class);
+
+    assertThat(result.getIssuesList()).extracting(Issue::getKey)
+      .containsExactlyInAnyOrder(project1Issue1.getKey(), project2Issue1.getKey())
+      .doesNotContain(project1Issue2.getKey(), project2Issue2.getKey());
+  }
+
+  @Test
+  public void search_by_application_and_project() throws Exception {
+    ComponentDto project1 = db.components().insertPublicProject();
+    ComponentDto project2 = db.components().insertPublicProject();
+    ComponentDto application = db.components().insertApplication(db.getDefaultOrganization());
+    db.components().insertComponents(newProjectCopy("PC1", project1, application));
+    db.components().insertComponents(newProjectCopy("PC2", project2, application));
+    IssueDto issue1 = db.issues().insertIssue(i -> i.setProject(project1));
+    IssueDto issue2 = db.issues().insertIssue(i -> i.setProject(project2));
+    userSession.registerComponents(application, project1, project2);
+    permissionIndexerTester.allowOnlyAnyone(project1);
+    permissionIndexerTester.allowOnlyAnyone(project2);
+    indexIssuesAndViews();
+
+    SearchWsResponse result = ws.newRequest()
+      .setParam(PARAM_COMPONENT_KEYS, application.getDbKey())
+      .setParam(PARAM_PROJECT_KEYS, project1.getDbKey())
+      .executeProtobuf(SearchWsResponse.class);
+
+    assertThat(result.getIssuesList()).extracting(Issue::getKey)
+      .containsExactlyInAnyOrder(issue1.getKey())
+      .doesNotContain(issue2.getKey());
+  }
+
+  @Test
+  public void search_by_application_and_project_and_leak() throws Exception {
+    Date now = new Date();
+    ComponentDto application = db.components().insertApplication(db.getDefaultOrganization());
+    // Project 1
+    ComponentDto project1 = db.components().insertPublicProject();
+    db.components().insertSnapshot(project1, s -> s.setPeriodDate(addDays(now, -14).getTime()));
+    db.components().insertComponents(newProjectCopy("PC1", project1, application));
+    IssueDto project1Issue1 = db.issues().insertIssue(i -> i.setProject(project1).setIssueCreationDate(addDays(now, -10)));
+    IssueDto project1Issue2 = db.issues().insertIssue(i -> i.setProject(project1).setIssueCreationDate(addDays(now, -20)));
+    // Project 2
+    ComponentDto project2 = db.components().insertPublicProject();
+    db.components().insertSnapshot(project2, s -> s.setPeriodDate(addDays(now, -25).getTime()));
+    db.components().insertComponents(newProjectCopy("PC2", project2, application));
+    IssueDto project2Issue1 = db.issues().insertIssue(i -> i.setProject(project2).setIssueCreationDate(addDays(now, -15)));
+    IssueDto project2Issue2 = db.issues().insertIssue(i -> i.setProject(project2).setIssueCreationDate(addDays(now, -30)));
+    // Permissions and index
+    userSession.registerComponents(application, project1, project2);
+    permissionIndexerTester.allowOnlyAnyone(project1);
+    permissionIndexerTester.allowOnlyAnyone(project2);
+    indexIssuesAndViews();
+
+    SearchWsResponse result = ws.newRequest()
+      .setParam(PARAM_COMPONENT_KEYS, application.getDbKey())
+      .setParam(PARAM_PROJECT_KEYS, project1.getDbKey())
+      .setParam(PARAM_SINCE_LEAK_PERIOD, "true")
+      .executeProtobuf(SearchWsResponse.class);
+
+    assertThat(result.getIssuesList()).extracting(Issue::getKey)
+      .containsExactlyInAnyOrder(project1Issue1.getKey())
+      .doesNotContain(project1Issue2.getKey(), project2Issue1.getKey(), project2Issue2.getKey());
+  }
+
+  @Test
+  public void search_by_application_and_by_leak_when_one_project_has_no_leak() throws Exception {
+    Date now = new Date();
+    ComponentDto application = db.components().insertApplication(db.getDefaultOrganization());
+    // Project 1
+    ComponentDto project1 = db.components().insertPublicProject();
+    db.components().insertSnapshot(project1, s -> s.setPeriodDate(addDays(now, -14).getTime()));
+    db.components().insertComponents(newProjectCopy("PC1", project1, application));
+    IssueDto project1Issue1 = db.issues().insertIssue(i -> i.setProject(project1).setIssueCreationDate(addDays(now, -10)));
+    IssueDto project1Issue2 = db.issues().insertIssue(i -> i.setProject(project1).setIssueCreationDate(addDays(now, -20)));
+    // Project 2, without leak => no issue form it should be returned
+    ComponentDto project2 = db.components().insertPublicProject();
+    db.components().insertSnapshot(project2, s -> s.setPeriodDate(null));
+    db.components().insertComponents(newProjectCopy("PC2", project2, application));
+    IssueDto project2Issue1 = db.issues().insertIssue(i -> i.setProject(project2).setIssueCreationDate(addDays(now, -15)));
+    IssueDto project2Issue2 = db.issues().insertIssue(i -> i.setProject(project2).setIssueCreationDate(addDays(now, -30)));
+    // Permissions and index
+    userSession.registerComponents(application, project1, project2);
+    permissionIndexerTester.allowOnlyAnyone(project1);
+    permissionIndexerTester.allowOnlyAnyone(project2);
+    indexIssuesAndViews();
+
+    SearchWsResponse result = ws.newRequest()
+      .setParam(PARAM_COMPONENT_KEYS, application.getDbKey())
+      .setParam(PARAM_SINCE_LEAK_PERIOD, "true")
+      .executeProtobuf(SearchWsResponse.class);
+
+    assertThat(result.getIssuesList()).extracting(Issue::getKey)
+      .containsExactlyInAnyOrder(project1Issue1.getKey())
+      .doesNotContain(project1Issue2.getKey(), project2Issue1.getKey(), project2Issue2.getKey());
   }
 
   private RuleDto newRule() {
@@ -583,7 +765,7 @@ public class SearchActionComponentsTest {
       .setName("Rule name")
       .setDescription("Rule desc")
       .setStatus(RuleStatus.READY);
-    dbTester.rules().insert(rule.getDefinition());
+    db.rules().insert(rule.getDefinition());
     session.commit();
     return rule;
   }
@@ -593,14 +775,14 @@ public class SearchActionComponentsTest {
   }
 
   private IssueDto insertIssue(IssueDto issue) {
-    db.issueDao().insert(session, issue);
+    dbClient.issueDao().insert(session, issue);
     session.commit();
     indexIssues();
     return issue;
   }
 
   private ComponentDto insertComponent(ComponentDto component) {
-    db.componentDao().insert(session, component);
+    dbClient.componentDao().insert(session, component);
     session.commit();
     return component;
   }
@@ -611,5 +793,10 @@ public class SearchActionComponentsTest {
 
   private void indexView(String viewUuid, List<String> projects) {
     viewIndexer.index(new ViewDoc().setUuid(viewUuid).setProjects(projects));
+  }
+
+  private void indexIssuesAndViews() {
+    issueIndexer.indexOnStartup(null);
+    viewIndexer.indexOnStartup(null);
   }
 }
