+++ server/sonar-server/src/test/resources/org/sonar/server/issue/ws/SearchActionTest/assignedToMe_facet_must_escape_login_of_authenticated_user.json
