+++ server/sonar-server/src/test/resources/org/sonar/server/issue/ws/SearchActionTest/display_deprecated_debt_fields.json
