+++ server/sonar-server/src/test/resources/org/sonar/server/issue/ws/SearchActionTest/display_facets.json
