+++ server/sonar-server/src/test/resources/org/sonar/server/issue/ws/SearchActionTest/filter_by_assigned_to_me.json
