+++ server/sonar-server/src/test/resources/org/sonar/server/issue/ws/SearchActionTest/hide_rules.json
