+++ server/sonar-server/src/test/resources/org/sonar/server/issue/ws/SearchActionTest/issue_on_removed_file.json
