+++ server/sonar-server/src/test/resources/org/sonar/server/issue/ws/SearchActionTest/issue_with_comment_hidden.json
